requirements.txt
alembic==1.12.0
fastapi==0.110.0
pydantic==2.3.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
uvicorn==0.28.0
fastapi-pagination==0.12.19
